Add remote calibration and motor mapping
Convert pin constants to uint8_t, add remote calibration/noise filtering and improved motor mapping. Changes include:
- bcapp/zeroremote.h: implement center calibration (setCenter), tolerance/statistics (Stat), sendCenter/sendXY, filtering of small changes and center handling, and updated loop logic.
- bcapp/zerocat.h: scale joystick inputs, compute radial component and drive motors via moveAnalog blending.
- bcstem/mcu.h: pin types changed to uint8_t.
- bcstem/motorset.h: add MIN_OUTPUT, PWM settings and remap analog output scaling to ensure motor responsiveness; use ledcAttach for PWM.
- bcstem/util.h: add Stat helper for sampling/derivation.
These changes improve responsiveness, reduce noise, and provide better PWM motor control.

+++ bcstem/mcu.h
@@ -31,8 +31,8 @@ GPIO3
 
   static const bool needDelayForSerialSetup = true;
   static const long bandRate = 115200;
-  static const int SDA = 8;
-  static const int SLC = 9;
+  static const uint8_t SDA = 8;
+  static const uint8_t SLC = 9;
 
   // MAC Address: b0:cb:d8:c6:52:04
 };
@@ -40,16 +40,16 @@ GPIO3
 struct ESP32S {
   static const bool needDelayForSerialSetup = false;
   static const long bandRate = 115200;
-  static const int SDA = 21;
-  static const int SLC = 22;
+  static const uint8_t SDA = 21;
+  static const uint8_t SLC = 22;
   // MAC Address: b0:cb:d8:c6:52:04
 };
 
 struct ESP32S3 {
   static const bool needDelayForSerialSetup = false;
   static const long bandRate = 115200;
-  static const int SDA = 8;
-  static const int SLC = 9;
+  static const uint8_t SDA = 8;
+  static const uint8_t SLC = 9;
   // MAC Address: b0:cb:d8:c6:52:04
 };
